Add files via upload
RateReadFile.m is just the function that reads the rateLibrary_pp.data and puts all the rates in a 28x7 matrix. 
ratefun.m separates each Rate parameter into a 28x1 matrix and computes each reaction's rate 
(note the resulting rate is in units of 10^10 (10 billion) and has a few 0.000 entries. When these entries are pulled out, the rate function will give the correct value. For example rate(1) is on the order of 10^-21, and appears as 0.000 in the rate function, but when displayed it shows the correct value)